Problems 67, 69 added. Problems 18, 71 edited.
p67 was already "solved" before (the exact same algorithm to solve it
was already used in a previous problem), and p18.py was edited to solve
both problems (p18 itself, and p67) through their corresponding data
files.

p69.py is in this repo symbolically, i.e., no computing effort at all
was needed to solve it.

p71.py was edited to comply with the style standard for this repo, as
well as to document it in order to understand it again.